lorgnette: Don't require platen for cli scans
lorgnette_cli assumes a default scan_source of Platen, which fails if a
scanner only has an ADF.  Switch the default to a new "Auto" value,
which tries the platen and falls back to the ADF if there is no flatbed
available.  If scan_source is explicitly specified, the previous
behavior of requiring a match is preserved.

BUG=b:261583360
TEST=lorgnette_cli scan on scanners with and without platen

Change-Id: I93b7e7c10833792b732dedd3f7e77b9f3d3cea43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5397805
Reviewed-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>