vm_tools: Integrate use flag for switching video decoder to libvda-vd
Today, if crosvm-virtio-video is set, then ARC-VM uses the host's Chrome
VDA mojo service to perform video decoding. It is desirable to migrate
from the VDA-based mojo service to a VD-based one. VDA is only used by
ARC, so removing the dependency unblocks removing the legacy VDA
ChromeOS decoder code.

This integrates the crosvm-virtio-video-vd USE flag to let a target
specify that it should use the libvda-vd decoder instead of the current
default libvda decoder. Individual targets will be modified to switch
USE crosvm-virtio-video-vd and migrate to the VD-based decoder in
stages.

This USE flag may not be long-lived. The goal of b:321171964 is to
switch the default to libvda-vd. When enough target are migrated to
libvda-vd, the default could be changed. Eventually a decoder type USE
flag could be removed.

If crosvm-virtio-video is not set, then crosvm-virtio-video-vd USE flag
contract is that it has no effect on the vm state. It is unnecessary to
raise an error.

BUG=b:321171964
TEST=cros_workon_make --test --board=${BOARD} vm_host_tools
TEST=Checked ARCVM args on builds with USE flag set or not set

Change-Id: I08caf67a39b57ac318e5bf361fc40610a8e02b6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5397803
Tested-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>