arc-setup: Check USE_ARCVM in arc-prepare-host-generated-dir
Currently, arc-prepare-host-generated-dir checks the IS_ARCVM env var
to see whether ARCVM is enabled on the board.
The env var corresponds to the enable-arcvm Chrome switch, which is set
in LoginManager's AddArcFlags() during the browser setup if ARCVM is
enabled on the board.
However, this initialization does not take place in guest sessions.
As a result, arc-prepare-host-generated-dir thinks it is a container
board, looks for /usr/share/arc instead of /usr/share/arcvm, and crashes
due to ENOENT.
Actually the indirection via the Chrome switch should just be redundant,
since we can directly check the board USE flag via USE_ARCVM.

BUG=b:306578991
TEST=Confirm that arc-prepare-host-generated-dir does not crash in guest
TEST=sessions.
TEST=$ grep -r IS_ARCVM arc/setup

Change-Id: I9c13388fb2841e01a42819891448af8a6d4591c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5400857
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Reviewed-by: Vaibhav Raheja <vraheja@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>