shill: Add SupportsConcurrency to WiFiPhy
Add a new private method SupportsConcurrency to WiFiPhy which determines
whether or not the Phy is able to support a set of interfaces
concurrently. This method replaces the behavior of WiFiPhy::CanUseIface.
CanUseIface was capable of very similar functionality, but it was
limited to concurrency with existing interfaces, and for upcoming
concurrency implementation, we need to be able to judge concurrency with
arbitrary hypothetical combinations of interfaces. SupportsConcurrency
is more condusive to this need.

SupportsConcurrency replaces WiFiPhy::SupportConcurrency since it
covers a superset of the functionality of that function.

BUG=b:328073403
TEST=FEATURES=test emerge-brya shill.

Change-Id: Ie7f329fddc90f709a944b58dcb0ef79de55c3527
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5502876
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>