shill: Add RequestNewIface to WiFiPhy
Add RequestNewIface to WiFiPhy, which allows callers to request a new
interface with a given type and priority. This function checks the
existing interfaces and their priorities and returns the set of
interfaces that should be destroyed in order to make room for the
requested interface. Returns a nullopt if the interface can't be
created at the given priority.

BUG=b:328073954
TEST=FEATURES=test emerge-brya shill

Change-Id: I0bc5a2cd02ce9c621b495858f34ed9e8346801c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5502877
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>