hwsec-host-utils: x-ver: add output options
The CL introduces two options for output in x-ver preparation.

1. --no-output
If specified, nothing would be output or uploaded. Useful when testing
the flow.

2. --output-tast
If specified, the output file would be put in the directory of x-ver
fixture data in platform/tast-tests. Useful when automating.

BUG=b:331504778
TEST=manual

Change-Id: Id7f9aebe5e6f026fea8318d04f0328455bc2d92b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5404654
Tested-by: Ching-Kang Yen <chingkang@chromium.org>
Reviewed-by: John L Chen <zuan@chromium.org>
Commit-Queue: Ching-Kang Yen <chingkang@chromium.org>