shill: Add Priority to WiFi device
Add a priority member to WiFi device. This value will be used for
concurrency conflict resolution. This value defaults to 4 on
construction of the device, as per go/cros-wifi-concurrency. A Manager
D-Bus property will be added in the future to allow API consumer to set
the priority of a WiFI device.

BUG=b:328074924
TEST=FEATURES=test emerge-brya shill

Change-Id: I158d12408cbed7a755030796574d1a07ba140945
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5502873
Reviewed-by: Andrzej Ostruszka <andrzejo@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>