power: Invoke a callback when the device key is restored
UserDataAuthClient handles the EvictedKeyRestored signal from
cryptohome. Currently the auth client decides which actions to
perform in response to the signal. That shouldn't really be
its responsibility and isn't practical as more actions are
needed, instead provide the auth client with a callback to
invoke when the signal is received. The callback is part of
the Daemon class, it currently just thaws the frozen
processes, as the auth client did.

BUG=b:329139212
TEST=USE=key_eviction FEATURES=test emerge-${BOARD} power_manager

Change-Id: I75fc54a92fc8b1e500f20071c177839587488968
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5460305
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>