diagnostics: Remove LedLitUpRoutineReplier in mojom
The `LedLitUpRoutineReplier` interface is no longer needed as the
interaction request is put into the waiting state. Also update the
comments of `LedLitUpRoutineArgument`.`

Notice that this change is not backward compatible. Other copies of the
Mojo definition (for example, the one in the chromium codebase) should
be updated together.

BUG=b:309781398
TEST=FEATURES=test emerge-volteer diagnostics
NO_IFTTT=No usage change in the CLI tool

Change-Id: I33cdfc7d44a66727f1cd7addf199178ee7abd3cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5397977
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Tested-by: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Kerker Yang <kerker@chromium.org>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>