patchpanel: add brillo-http PrepareTagSocket API
Add a PrepareTagSocket() call to request the application of a traffic
annotation to a socket owned by a libbrillo-http Transport. A callback
is registered to the Transport, it will be called when the socket is
created but before the first connection. Then a TagSocket call is done
to apply the socket annotation.

As a synchronous TagSocket call may cause a deadlock between the caller and
the callee, it is asynchronous for now. It may lead to a situation where
the socket start being used before the TagSocket call is processed. It
will be addressed in a subsequent CL.

BUG=b:331620358
TEST=FEATURES="test" emerge-$BOARD patchpanel patchpanel-client

Change-Id: Idcb79c0ed6b1956342bf4582de8dbead412aa0eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5407696
Commit-Queue: Damien Dejean <damiendejean@google.com>
Tested-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Cassie Wang <cassiewang@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>