shill: implement "http-only" for Service.CheckPortal
This patch adds support in Service CheckPortal for the "http-only" mode
which enabels portal detection with HTTP probes only and skips Internet
connectivity checks with HTTPS probes. This patch only connects the
CheckPortal property for "http-only" to the previously introduced
http_only option of PortalDetector.

BUG=b:318370676
TEST=FEATURES=test emerge-$BOARD shill. Manual: flashed shill, set
CheckPortal to "http-only" on GoogleGuest, verified in /var/log/net.log
that a single HTTP request is sent. Verified that "CheckPortal" is set
to "http-only" in stored Service properties for GoogleGuest in
/var/cache/shill/default.profile.

Change-Id: Ib8231a95acfe20ac467c70f7d5c08f50400523ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5409955
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>