shill: Add concurrency helper types to WiFiPhy
This change adds some helpful types to WiFiPhy that will be used in
concurrency conflict resolution. The main purpose of these types is so
that we can keep the comparison logic isolated to a single location.

ConcurrentIface: A single interface to be considered during concurrency
conflict resolution. Contains data about the interface which is relevant
to concurrency.

RemovalCandidate: A set of ConcurrentIfaces sorted by priority.
Represents a set of interfaces that we'll consider removing while
resolving a concurrency conflict.

RemovalCandidateSet: A set of RemovalCandidates sorted by their
preferability. Keeping these sorted allows us to consider
RemovalCandidates one at a time and take the first viable option we
encounter.

BUG=b:328073954
TEST=FEATURES=test emerge-brya shill

Change-Id: I7a3d27a8a069e522c7a200fe8b5bf58c998b1cdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5502875
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>