patchpanel: Remove unused MOCK_METHOD in mock_patchpanel_proxy
Provide a stub implementation and build the mock class on top of that.

Compiling a mock class with a lot of mock methods is expensive and thus
we want to remove the MOCK_METHOD definition if it's not required in the
test.

This patch can reduce the patchpanel-client compile time by ~4 seconds.

Also change the fuzzer to use the stub implementation.

BUG=b:332466760
TEST=FEATURES="test" emerge-$BOARD patchpanel-client

Change-Id: I5293ebc06d7958f2313b0f63ce45629e7c7bf46b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5418273
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>