cryptohome: Add support for recovery-prepare to the cli
This adds a prepare_recovery_auth_factor action that's basically the
same as the get_recovery_request action today, but using the
PrepareAuthFactor form of the operation.

This could instead be added by extending the more generic
prepare_auth_factor command but in practice that makes that command a
bit of a mess since you either have to allow some weird combinations or
you have to add special handling when the command is invoked for
recovery factors. Making a dedicated action is a little simpler, and
actually a bit more user friendly.

The implementation doesn't allow you to set all of the tokens in the
prepare input that would actually be set on a real recovery. However,
this is just mirroring what the existing get_recovery_request already
does and it has the same limitation. Presumably because these actions
were only intended for use with test mockups where those tokens are not
actually used.

BUG=b:269619280
TEST=FEATURES=test emerge-$BOARD cryptohome-dev-utils

Change-Id: I2d795dd6a369498fe1bd32d747d61319bb14f750
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5413874
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>