permission_broker: use legacy_usb_passthrough USE flag
For permissive USB passthrough (see go/lofn) we will need to default to
legacy behavior for platforms which do not (and likely will not) have
USB annotations. These 3 platforms are also excluded in
https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/platform/tast-tests/src/go.chromium.org/tast-tests/cros/remote/bundles/cros/firmware/usb_annotations.go;l=87;drc=b831391f486f3f8c9a7306487f6ec36862d1be85,
and we therefore feel that since this test is passing for all platforms
this should be a pretty 'authoritative' list.

We use the value of the USE flag which defaults to false and is set to
true for the appropriate board overlays.

BUG=b:267951284
TEST=build

Cq-Depend: 5426651
Cq-Depend: 5426857
Change-Id: I0a4bc15885ec555358408a61e1c3a0f0371b4e5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5415717
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: maciek swiech <drmasquatch@google.com>
Tested-by: maciek swiech <drmasquatch@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>