fpstudy: Allow the result to be used as a general error
This result/error that is reported to the user website is currently
not being used. It was originally designed to provide feedback about
the current capture (ex. low coverage, move finger). If any result
other than "ok" was sent back to the user page, the error message
would be shown to the user and the capture would be retried.

This does not change the behavior of this error mechanism, but merely
allows it to present free-form alternative error messages.

BUG=b:324677012
TEST=./host-run.sh

Change-Id: I83de34fb3971659ca0a65a9cd22bc9152b06b524
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5417576
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>