patchpanel: Add protobuf headers to PCH
This is a rework of https://crrev.com/c/5239659.

The generated protobuf header is included by a lot of translation units
in patchpanel. To improve the incremental build speed, including it in
PCH can be helpful, but seems that sometimes the build system cannot
recompile PCH correctly in some cases (the reason is unclear, probably
when proto and patchpanel are changed at the same time).

Instead of including the generated header, this patch includes the
protobuf library files which are included by that header instead. These
headers should be much more stable and it usually won't be changed
during the development.

BUG=None
TEST=FEATURES="test" emerge-$BOARD patchpanel

Change-Id: I173cb81fc55390e89e8f6655107e0460e82d750c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5418735
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>