crosh: shell: write eventfd directly from SIGWINCH handler
The write() function is one of the async-signal-safe functions, per
<https://man7.org/linux/man-pages/man7/signal-safety.7.html>, so we
should be able to write to the resize eventfd directly from the signal
handler rather than setting a flag to be checked by another thread.

This avoids the need for an extra thread that wakes up every 100ms.

BUG=b:330373462
TEST=run crosh shell --isolated and resize

Change-Id: Ie3ddcc2dcc34758ef5735a60f7928a5e21c48612
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5420982
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ben Scarlato <akhna@google.com>