libchromeos-ui: Also use VulkanFromANGLE with vulkan_chrome
The Finch experiment for Vulkan enables three flags: Vulkan,
DefaultANGLEVulkan, and VulkanFromANGLE [1]. The vulkan_chrome USE
flag is meant to be switched on right before the Vulkan Finch
experiment rolls out to 100% stable, so the sets of Chrome flags
should be the same.

Currently, vulkan_chrome is missing VulkanFromANGLE, so this CL
adds it to match the Finch.

[1] http://google3/googledata/googleclient/chrome/finch/gcl_studies/Vulkan.gcl;l=94;rcl=621217707

BUG=b:292249282
Test=Flash to device and verify args in chrome://gpu

Change-Id: Iae42ff19ab508512ffd86873a0701280c2ec0ec6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5421614
Commit-Queue: Brian Ho <hob@chromium.org>
Tested-by: Brian Ho <hob@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>