cryptohome: Relocate auth_input_utils
This is not particularly important, just a minor cleanup of some code I
was recently working on for adding recovery+prepare support.

The intent of this refactor is to address a few housekeeping things:
* reduce the amount of random files in the top level cryptohome dir
* get rid of files with unhelpful "utils" names
* separate out the PrepareOutput utilities
It's still technically a little misleading to put the prepare output
utility in "auth_io" the AuthInput and PrepareInput and PrepareOutput
all feel related enough that making two separate directories seemed a
little overkill.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I1b35fb261bd2049d2598ad0ddc1c67209af369e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5421615
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Zi Lin <lziest@google.com>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>