lorgnette: Improve activity tracking in discovery sessions
lorgnette exits 2s after the last activity.  In the case of discovery
sessions, it was tracking the start time of a session instead of the
last activity, which could cause it to exit too quickly on slow devices.

Fix this with two related changes:
  * Track the last activity on active sessions instead of the start
    time.
  * Track session termination as an activity update so that the timer
    doesn't expire immediately after the last session exits.

BUG=b:331069087, b:276909624
TEST=20 iterations of tast documentscanapi.Scan on pazquel

Change-Id: I7dca395879e0c47a72a92be6418b1a501565c3ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5421440
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>