crash: Change GenericFailureCollector to return CrashCollectionStatus
Change GenericFailureCollector to return CrashCollectionStatus from its
collection functions.

Unlike some of the other changes for this bug, this one does introduce
one functional change: errors in the generic failure collector will
cause crash_reporter to return exit code 1 (failure) instead of 0
(success). Before, the code always return 0 even when it failed to write
an error report because anomaly_detector would CHECK-fail on an exit
code of 1. That's changed in the parent CL, so that
GenericFailureCollector is free to return a sensible return value.

As part of this CL, changed CrashCollector::GetLogContents and related
functions to return CrashCollectionStatus. Previously, I had thought the
return value of CrashCollector::GetLogContents only controlled if the
logs were attached to the crash reports, so I left it as a boolean. But
here, a failure in GetLogContents actually fails the crash report
completely, so I want a better idea of what exactly went wrong.

BUG=b:177552411
TEST=Together with parent CL, created out-of-spool-space condition,
ensured no crashes when new anomalies found.
TEST=Ensured service failures created new error reports

Change-Id: I4ad748ad5b41ed11addab4c317cf818ce4bf24a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5421443
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>