debugd: handle resize eventfd in main event loop
We already have a select()-based event loop that is handling the copying
of data to and from the pty, which can be extended to handle the resize
eventfd as well, rather than launching a separate thread.

While we're here, refactor the common TIOCGWINSZ + TIOCSWINSZ sequence
into a helper function instead of repeating it.

BUG=b:326266002
TEST=shell --isolated, run vim, and resize the window

Change-Id: Ie33d168915754d9229024290dc8d4d725e8e7204
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5425498
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>