typecd: Fix DPAM cable notification for emarked USB 2.0 only cables
Currently, typecd will only send the DPAM cable notification on boards
without AP driven mode when the port manager concludes that DP alt mode
can be entered and the cable is invalid. But, typecd no longer supports
DPAM entry with emarked USB 2.0 only cables. So, the notification will
not be sent when a display is connected with an emarked USB 2.0 cable,
even though the cable is the reason that DPAM cannot be entered. This
updates the port manager to send the DPAM cable notification when the
invalid cable flag is set. This will only ever be set when the partner
supports DPAM.

BUG=b:332957452
TEST=FEATURES=test emerge-${BOARD} chromeos-base/typecd, cable
notification correctly sent on board without AP driven mode entry when
connecting a display with an emarked USB 2.0 cable.

Change-Id: I3d8276cdaf59e03aba9b5fa6ab03729ccd9a2795
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5426868
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>
Tested-by: Jameson Thies <jthies@google.com>