login: Use a re-enrollment key as the state key if it is provided
See go/fre-on-flex for rationale and design.

If machine info params contain a re-enrollment_key (a hex string of
64 characters, for a 256 bits key), use this as the state key, and
only return that state key without quantization.

On Flex, we will provide this by calling the flex_id_tool with an
appropriate argument to supply us with the so-called Flex state key.
This will be done in ui-collect-machine-info.

BUG=b:332709076
TEST=FEATURES=test emerge-${BOARD} chromeos-login

Change-Id: I4fa3e60aa546e92842c591b8bf85d98115e6a9c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5425497
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Yves Arrouye <drcrash@chromium.org>