shill: vpn: Introduce VPNEndReason to describe driver failures (part 1)
Prior to this patch, Service::ConnectFailure is used to described why
the VPN connection is lost or the connection cannot be established.
Since that enum is used by all technologies in shill, most of the fields
won't be reported by VPN services, and the existing fields are not
detailed enough to describe a VPN failure.

This patch introduces the VPNEndReason enum for this purpose, it can be
mapped into the ConnectFailure when VPNService needs to passing it up
to Service, and will also be used in VPN metrics later.

This patch also changes the OpenVPN, ThirdPartyVPN, and WireGuard to
use enum. Note that:
- Some of the reported service failures are changed by this patch.
- Eventually VPNDriver class will pass this enum to VPNService directly.
  It's left as-is now to avoid too many changes in the same patch.

BUG=b:331743444
BUG=b:231910338
TEST=FEATURES="test" emerge-$BOARD shill

Change-Id: Ifc279c558c5cb2d48455310c6719400c5f3db21c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5426254
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>