shill: service: remove ConnectState::kPortalSuspected
This Service connection state was effectively folded into the
kRedirectFound state as part of crrev/c/5384682 and there is no UX
distinction anymore. Since the portal detection state machine in the
platform already behaves identically for
ValidationState::kPortalRedirect and ValidationState::kPortalSuspected,
it is not necessary to distinguish them in the Service. For metrics
purposes, NetworkMonitor will keep the distinction internally to shill.

BUG=b:333634158
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I403fa76126da1a74f22709d92a7b95aa34f6ce27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5442413
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>