init: startup: Drop EFI system partition mount
There are a couple problems with mounting the ESP during init:

First, when installing the OS, chromeos-install will try to DD the
source ESP to the destination ESP. It's not a good idea to do that while
the ESP is mounted rw.

Second, fsck thinks the ESP is dirty after unmounting it. The vfat
driver marks the filesystem dirty as soon as it is mounted RW (see
https://unix.stackexchange.com/a/230185), but that bit should be cleared
on unmount. For some currently-unknown reason, when the ESP is mounted
by chromeos-init it stays dirty after a manual unmount.

The ESP mount was added to support firmware updates on ChromeOS Flex,
but that feature is not yet fully implemented anyway, so it should be
safe to disable the ESP mount until the problem can be fully
investigated.

BUG=b:332941579
TEST=FEATURES=test emerge-reven chromeos-init
TEST=cros build-packages && cros build-image -r test
TEST=Run in a UEFI VM, verify /efi is not mounted
TEST=Install image, verify no warning about ESP corruption

Change-Id: Ida6e9b007139442a7295e1789f30d31a7cc9df43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5426533
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Alexis Savery <asavery@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>