arc/container: check for CTS mode conservatively
When applying a hack for running CTS tests, the check would sometimes
give a false positive, which would cause the hack to be applied during
normal use. This is because the udev rules were always evaluating to
false for the "unknown" case. So we change != to == such that "unknown"
causes non-test mode to activate.

TEST=run cts -m CtsHardwareTestCases -t android.hardware.input.cts.tests.MicrosoftXbox2020Test
TEST=run cts -m CtsHardwareTestCases -t android.hardware.input.cts.tests.SonyDualshock3UsbTest
TEST=run cts -m CtsInitTestCases -t keychords#init_inotify
TEST=run cts -m CtsInitTestCases -t keychords#key
BUG=b:319553374

Change-Id: I6e820bb9aa162cda95af3708342b88895c2804ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5430633
Reviewed-by: Yury Khmel <khmel@chromium.org>
Tested-by: Matthew DeVore <matvore@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>