crash: Modify ParseCrashLog to parse ANR reports properly
Current implementation stops reading the report contents for headers
after a new line. The "Subject" header was moved to the data file for
ANR reports instead of staying with other headers in ag/14487423.

This "Subject" is being used for ANR classification and not being able
to upload it to the crash reports makes ANR reports very generic.

This CL modifies ParseCrashLog to make it keep trying to find Subject
even after a new line.

BUG=b:293810347
TEST=Manually triggering ANR and checking the crash report
http://crash-staging/browse?reportid=57b1fe1713ee55a3&index=0#0

Change-Id: I985726a3121b249ea214e98f2c7287e3a28cd5d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5436533
Tested-by: Bryan Yu <yuholong@google.com>
Commit-Queue: Bryan Yu <yuholong@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Daichi Hirono <hirono@chromium.org>