cryptohome: add a feature flag to gate legacy fp migration.
FpMigrationUtility takes a pointer to featured client so
it would give authortative runtime migration rollout number.

Additionally, FpMigrationUtility provides a new boolean method
NeedsMigration, which is used by AuthSession to decide
whether to carry out the whole migration flow.

At the end of the migration flow, the rollout number is persisted
to USS. And a unit test is updated to check that on the happy path.

BUG=b:327048355
TEST=CQ;modified unit tests.

Change-Id: I8d9fd640c1b9d439c3e906b2cce1ba3c8d3c1a0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5437054
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Zi Lin <lziest@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>