shill: vpn: Report end reason metrics
This patch adds the following metrics:
- ${VPNType}.ConnectFailureReason: Reported on failure of connecting to
  a VPN service.
- ${VPNType}.ConnectionLostReason: Reported on a connection lost event
  of a established VPN connection.

VPNEndReason will be mapped to the values of these two events (plus the
"reconnect" for the second event which is not included in VPNEndReason).

The current EndReason metrics in L2TPIPSecDriver and IKEv2 driver will
be removed in the next milestone for comparison (there won't be name
conflicts).

BUG=b:331743444
BUG=b:231910338
TEST=FEATURES="test" emerge-$BOARD shill

Change-Id: I3ce47f777d581658cf6297e5f3f0d49e4340cb9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5437638
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>