vm_tools: Stop taking metrics from StopAllVm
This is preparation to remove D-Bus thread in vm_concierge.
As a part of the effort, we're going to get rid of thread-based
VM shutdown. That makes it difficult to take each VM's metrics,
so this CL removes it, as VM team says it's no longer critical.

BUG=b:289932268
TEST=Tryjob.

Change-Id: I829614184067df7de5f14a23a53c3e1b810c191c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5440632
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Ryuichiro Chiba <chibar@chromium.org>