crash: Increase the tethering failure weight to capture logs
With the current weight and traffic expectations, it is very unlikely
that we'll ever be able to capture logs for these failures. With the new
weight proposed here, we should be able to capture failures and still be
within the daily limit of 10k reports.

BUG=b:333546657
TEST=CQ

Disallow-Recycled-Builds: test-failures
Change-Id: Id04482c150c8535267c45f642504856c4ec34626
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5439007
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Nikhil Nayunigari <nikhilcn@google.com>
Commit-Queue: Nikhil Nayunigari <nikhilcn@google.com>