cryptohome: Add ERROR log needed for crash reporting
This was a bit of a regression from the relocation of the recovery
request generation to the Prepare flow. The new code doesn't emit this
ERROR which it seems the crash reporter is depending on for synthesizing
a "crash report" on recovery request generation failures.

This regression doesn't actually break the recovery flow itself, but
losing the reports obviously impacts some of the failure reporting.

BUG=b:333446742
TEST=unit tests

Change-Id: I1f127a0827bffa8dc364d85672572f720a978729
Disallow-Recycled-Builds: hana-cq
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5444763
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>