init: startup: Remove ' in ""
Using `` inside "" was a reminiscence of the script era, to make sure
the string was sent as a single argument. This is not needed anymore
since brillo::Process take care of it.

BUG=b:172210918
TEST=Create a directory `/mnt/stateful_partition/factory_install_reset`,
unexpected, as we expect a file, observe clobber.
Check /var/log/clobber.log output is expected:
'''
2024/04/10 18:36:37 UTC Powerwash initiated by Reset file presence, but invalid
'''

Change-Id: I18ab51faeaae1610b306530613bda61088f5fc65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5444766
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>