shill: portal: use "http-only" mode on networks with proxy configs
When a network has a per-Service proxy configuration, shill needs to
conservatively assume that the network may be some on-prem network
configured manually by the user (guest Enterprise WiFi, etc). This
requires to skip network validation and automatically classify the
network as online.

In most cases, the proxy configuration is not related to the on-prem
requirements of an unanounced Enterprise network and there is no local
proxy to go through to reach the internet. The typical example is when
the user configures a remote web proxy for navigation purposes.

To still support portal detection in these cases, the Service now
configures the validation mode of the Network to "http-only". This
enables detecting captive portal redirection with HTTP probes but
ensures that the Service will still transitions to the "online" state
when there is no portal.

BUG=b:207657239
TEST=FEATURES=test emerge-$BOARD shill. Manual: flashed device,
configured fake PAC URL configuration for GoogleGuest in the UI,
verified that network validation with probes is done with HTTP only.

Change-Id: I02f2fba4cb668b8bfa54dd84ae77cfb52aadad53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5446564
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>