init: upstart: See feature level at boot
boot-splash was seeding feature_level for the rest of the OS until
crrev.com/c/5080768, where it is not done if the firmware takes care of
the splash screen.

Add a new job that start early at pre-startup, after splash screen,
and seed the VPD if necessary - the first boot after update.

BUG=b:333307104
TEST=Check VPD is set even when removed from the boot-splash conf file.

Change-Id: I12c5748ae036a1f1a11dad3b95b38ea60b1d3904
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5446441
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>