mojo_service_manager: Fetch username to ProcessIdentity
This is a part of the security context migration (b/333324147).

Add a new field "username" to ProcessIdentity. This field tells clients
what user service manager got by the uid. Should be useful when clients
don't have access to the host user list.

Username is fetched when a socket connection is accepted by the service
manager. We already fetch the uid so we just need to get the username by
the uid.

Also add tests to verify this change.

BUG=b:333309396
TEST=FEATURES=test emerge-zork mojo_service_manager

Change-Id: I7d05b60e2160b4eaba32564ede497d703b5c64ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5439141
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Ben Scarlato <akhna@google.com>
Tested-by: Chung-sheng Wu <chungsheng@google.com>