login: Collect scheduler stats when browser is not responding
Collect statistics from the system scheduler for the browser process
when a ping response was not received. For the first timeout, print the
scheduler statistics at the time the ping was sent to see if and how
much time the browser was scheduled on the CPU.

BUG=b:333318113
TEST=pkill -SIGSTOP chrome; wait for some time between 6 and 60 sec;
Check that the logs contain statistics from the system scheduler for the
chrome process.

Change-Id: I1554e8619eb46ada4f47f29c06b8be588be34667
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5444625
Tested-by: Marek Maślanka <mmaslanka@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Marek Maślanka <mmaslanka@google.com>