camera: Add HW_VIDEO_ENCODER usage to FramingSM full frame stream
In FramingSM, set GRALLOC_USAGE_HW_VIDEO_ENCODER usage to full frame
stream if this usage is set in YUV stream requested from the client.
This is to solve the issue in CCA video mode in devices with camx HAL
that requires HW_VIDEO_ENCODER usage to be the same in all YUV streams.

BUG=b:333679213
TEST=tast run $strongbad camera.CCAUIDigitalZoom.video

Change-Id: I067c46517bda61d971b685b6b4cedd730db16c3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5446773
Commit-Queue: Kam Kwankajornkiet <kamchonlathorn@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Tested-by: Kam Kwankajornkiet <kamchonlathorn@chromium.org>